Get records in past year
